Implement a new algorithm for the first task
As for now, the first task can be solved using two algorithms:
* optimized version of the TLU-base algorithm
* a special algorithm, which makes the matrix lower-triangular

This commit adds all the necessary implementations. The first report is
updated accordingly: it now includes comparison of two algorithms and
some new testing data.

This commit includes also some refactoring changes. For example, the
project now has a special Random function, used inside the tests, and
DataOutput class, which can print the testing data to the files (this
data then can be used in the reports).